Main file with some changes
Here are changes to pushDataToDb mainly, I also added the attributes of journalQuery and categoryQuery to BasicQueryEngine and initialised them (I think we only need to as much, but let me know). Uploaded queries (until addJournalHandler included) -  they supposedly work (I tried them with Blazegraph and sqlite from bash), the only problem I'm facing right now is about the RDF relations, while processing some queries I was having the problem of double entries (false and true) for both the DOAJ and APC - I tried figuring out what was going on and I've tried a bunch of things but ate I think it has to do with the graph itself but I'll try to check with Giulia Renda (possibly) next week.