FINALLY ADD THOSE GODDAM ID METHODS
Thanks to burning my brain cells, I can say that BOTH the getById and getEntityById methods are finally made to a convincing level, meeting the requirements of the project specifications.

The getEntityById method is a much better way of creating instances of Journal, Area, and Category objects. We can apply it to all parts of the BasicQueryEngine. This also means that the code is much less error-prone and we don't have to write the same thing a bunch of times...