Refine the entity creation process
Each category in a journal now only contains one quartile. The quartile featured in each category is the one that is linked to each entry in the relational database. 

I have also made other changes to comply better with the UML diagram. Basic functions that return the attributes of entities have been checked to ensure that the output is appropriate, and they have also been used in the testing process to ensure that expected and actual outputs match with each other (i.e. any given journal has the correct attributes, but also the correct linked categories (and by extension quartiles) and areas.

Minor changes have been made to SPARQL queries that previously weren't working. The journals returned from each of these queries now have the column 'journal-ids' as a label instead of 'o'. While it is inefficient to decentralise this process and do it in every function, Peroni may penalise us if a journal returned in a dataframe is not compliant with what is expected, even with as minuscule details as what a label might be. I don't want to take any risks here. Note that a mark can be lost simply because a function returns output that is inappropriate and/or unexpected...